Code as revised at San Diego County Fair 2015
Reduce maximum allowed notes from 30 to 24.

Send a constant stream of redundant Note Off commands (one every 5ms)
to notes that aren't supposed to be playing. This is supposed to fix
any problem manifesting as stuck notes, whether due to a bug in the
MTP8 or an error in the controller code.

Changed the note counting scheme and the function of requestOn and
requestOff. Now those functions handle note counting. This fixes a
problem where note counting caused us to discard a Note On, but not the
corresponding Note Off, which led to the note count being decremented
spuriously, and thus to excess notes being allowed to play, and thus to
blown fuses.

Changed external MIDI to map to rank 1 instead of rank 0.

Invoke allOff() when the external MIDI enable button is turned off, to
clear out notes currently playing when that happened.