Change to limit note count separately by MTP
In addition to overall note counting, we now count (and limit) the
notes on each MTP8 board. This better reflects the actual current
handling limitations of the hardware, but it requires an additional
setup step to include a table of note-to-MTP assignments in the code
for each deployment.

A Python script deploy.py is run on a development computer to create
Arduino code initializing a table describing which pipes are connected
to which MTP8.

Untested.